// frontend/src/app/core/api/model/notificationDTO.ts
/**
 * Rev-PasswordManager API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface NotificationDTO { 
    id?: number;
    notificationType?: string;
    title?: string;
    message?: string;
    createdAt?: string;
    read?: boolean;
}